# api_usuarios_dsw/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "description": "API de autenticación y gestión de usuarios",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5501/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/registrar": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "24031ad56fbf4bc95a6316e93af6504b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan García"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "MiPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado con éxito"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Juan García"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "juan@email.com"
                                                },
                                                "rol": {
                                                    "type": "string",
                                                    "example": "Cliente"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos o email ya en uso"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@habita.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Admin1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "rol": {
                                            "type": "string",
                                            "example": "Administrador"
                                        },
                                        "abilities": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "muebles.crear"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/perfil": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Ver perfil del usuario autenticado",
                "operationId": "59ad886d5dbee6b17dac06a7624efc70",
                "responses": {
                    "200": {
                        "description": "Datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Juan García"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "juan@email.com"
                                        },
                                        "rol": {
                                            "type": "string",
                                            "example": "Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Introduce el token Bearer obtenido al hacer login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        }
    ]
}